uhyve_interface

Enum HypercallAddress

#[non_exhaustive]

#[repr(u16)]
pub enum HypercallAddress {

    FileWrite = 1_024,
    FileOpen = 1_088,
    FileClose = 1_152,
    FileRead = 1_280,
    Exit = 1_344,
    FileLseek = 1_408,
    Netwrite = 1_600,
    Netread = 1_664,
    Netstat = 1_792,
    Cmdsize = 1_856,
    Cmdval = 1_920,
    Uart = 2_048,
    FileUnlink = 2_112,
    SerialBufferWrite = 2_176,

}

Enum containing all valid port mappings for hypercalls.

The discriminants of this enum are the respective ports, so one can get the code by calling e.g., HypercallPorts::FileWrite as u16.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

FileWrite = 1_024

Port address = 0x400

FileOpen = 1_088

Port address = 0x440

FileClose = 1_152

Port address = 0x480

FileRead = 1_280

Port address = 0x500

Exit = 1_344

Port address = 0x540

FileLseek = 1_408

Port address = 0x580

Netwrite = 1_600

๐Deprecated: was never really in use

Port address = 0x640

Netread = 1_664

๐Deprecated: was never really in use

Port address = 0x680

Netstat = 1_792

๐Deprecated: was never really in use

Port address = 0x700

Cmdsize = 1_856

Port address = 0x740

Cmdval = 1_920

Port address = 0x780

Uart = 2_048

Port address = 0x800

FileUnlink = 2_112

Port address = 0x840

SerialBufferWrite = 2_176

Port address = 0x880