uhyve_interface

Module parameters

Parameters for hypercalls.

Structs

CloseParams

Parameters for a FileClose hypercall.

CmdsizeParams

Parameters for a Cmdsize hypercall which provides the lengths of the items in the argument end environment vector.

CmdvalParams

Parameters for a Cmdval hypercall, which copies the arguments end environment of the application into the VM’s memory.

ExitParams

Parameters for a Exit hypercall.

LseekParams

Parameters for a FileLseek hypercall

OpenParams

Parameters for a FileOpen hypercall.

ReadParams

Parameters for a FileRead hypercall.

SerialWriteBufferParams

Parameters for a SerialWriteBuffer hypercall.

UnlinkParams

Parameters for a FileUnlink hypercall.

WriteParams

Parameters for a FileWrite hypercall.

Constants

ALLOWED_OPEN_FLAGS

EBADF

Bad file number

EFAULT

Bad address

EINVAL

Invalid argument

ENOENT

No such file or directory

EROFS

Read-only file system

O_APPEND

O_CREAT

O_DIRECTORY

O_EXCL

O_RDONLY

O_RDWR

O_TRUNC

O_WRONLY

SEEK_CUR

The file offset is set to its current location plus offset bytes.

SEEK_END

The file offset is set to the size of the file plus offset bytes.

SEEK_SET

The file offset is set to offset bytes.